Move viewing service into the core folder
