+++ src/app/core/viewing.service.spec.ts
