+++ src/app/core/viewing.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { MsgIoService } from '../../core/msg-io.service';
-import { Fonts, VirtrolioMessage } from '../../shared/interfaces';
-import { FontService } from '../../core/font.service';
+import { MsgIoService } from './msg-io.service';
+import { Fonts, VirtrolioMessage } from '../shared/interfaces';
+import { FontService } from './font.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ViewportScroller } from '@angular/common';
 import { ToastrService } from 'ngx-toastr';
-import { AuthService } from '../../core/auth.service';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root'
