+++ src/app/pages/viewing/viewing.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { ViewingService } from './viewing.service';
+import { ViewingService } from '../../core/viewing.service';
 import { firestore } from 'firebase/app';
 import { VirtrolioMessage } from '../../shared/interfaces';
 import { Title } from '@angular/platform-browser';
